# OCR-Spring-Backend-main/swagger/ocr-api.yaml
openapi: 3.0.0
info:
  title: OCR API
  description: API for managing OCR apps
  version: 1.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://localhost
...
paths:
...
  /api/user/entry/add/{id}:
    post:
      tags:
        - Entry
      summary: Add a telecunsultation entry
      description: >
        This endpoint adds a Entry. If the patient is not found, it returns an error message.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The patient id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                  start_time:
                    type: string
                    format: date-time
                  end_time:
                    type: string
                    format: date-time
                  complaint:
                    type: string
                  findings:
                    type: string
                  current_habits:
                    type: string  
      responses:
        "200":
          description: telecunsultation entry created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeleConEntry" 
        "401":
          description: Unauthorized access
        "404" :
          description: Patient not found
        "500":
          description: Internal server error    
  /api/user/entry/get:
    get:
      tags:
        - Entry
      summary: get all entries added by user
      description: >
        This endpoint returns a list of all teleconsultation entries added by the user. It supports pagination and search.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: page number for pagination
        - in: query
          name: filter
          schema:
            type: string
            enum: [created Date,updated Date,Assigned,Unassigned,Reviewed,Unreviewed]
          description: filter for entries
      responses:
        "200":
          description: A list of telecunsultation entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeleConEntry"
        "401":
          description: Unauthorized access
        "500":  
          description: Internal server error
  /api/user/entry/get/patient/{id}:
    get:
      tags:
        - Entry
      summary: Get patient's entries
      description: >
        This endpoint returns a list of all teleconsultation entries for a specific patient. It supports pagination and search.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The patient Id
        - in: query
          name: page
          schema:
            type: integer
          description: page number for pagination
        - in: query
          name: filter
          schema:
            type: string
            enum: [created Date,updated Date]
          description: filter for entries
      responses:
        "200":
          description: A list of telecunsultation entries for the patient
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeleConEntry"
        "401":
          description: Unauthorized access
        "500":
          description: Internal server error
  /api/user/entry/shared/patient/{id}:
    get:
      tags:
        - Entry
      summary: Get shared patient's entries(view only)
      description: >
        This endpoint returns a list of all shared teleconsultation entries for a specific patient. It supports pagination and search.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The patient Id
        - in: query
          name: page
          schema:
            type: integer
          description: page number for pagination
        - in: query
          name: filter
          schema:
            type: string
            enum: [created Date,updated Date]
          description: filter for entries
      responses:
        "200":
          description: A list of telecunsultation entries for the patient
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeleConEntry"
        "401":
          description: Unauthorized access
        "500":
          description: Internal server error
  /api/user/entry/get/{id}:
    get:
      tags:
        - Entry
      summary: get one entry details added by users
      description: >
        This endpoint returns the details of a Entry. If the entry is not found, it returns an error message.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The entry Id
       
      responses:
        "200":
          description: telecunsultation entry details
          content:
            application/json:
              schema:
               $ref: "#/components/schemas/TeleConEntry"
        "401":
          description: Unauthorized access
        "500":
          description: Internal server error
  /api/user/entry/count/newreviews:
    get:
      tags: 
        - Entry
      summary: get count of new reviews
      description: >
        This endpoint returns the count of new reviews. If no reviews are found, it returns an error message.
      responses:
        "200":
          description: count of new reviews
          content:
            application/json:
              schema:
                type: integer
                example: 5
        "401":
          description: Unauthorized access
        "500":
          description: Internal server error
  /api/user/entry/count/newentries:
    get:
      tags: 
        - Entry 
      summary: get unreviewed entry count
      description: >
        This endpoint returns the count of unreviewed entries. If no unreviewed entries are found, it returns an error message.
      responses:
        "200":
          description: count of unreviewed entries
          content:
            application/json:
              schema:
                type: integer
                example: 4
        "401":
          description: Unauthorized access
        "500":
          description: Internal server error
  /api/user/entry/reviewer/add/{id}:
    post:
      tags: 
        - Entry
      summary: Add a reviewer by user
      description: >
        This endpoint adds a reviewer to an entry. If the entry is not found, it returns an error message.
      parameters:
        - in: path
          name: Id
          schema:
            type: string
          required: true
          description: The entry Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                  reviewer_id:
                    type: string
                    example: 5f7b1b3b7f3b9b0017b3b9b0
      responses:
        "200":
          description: reviewer added successfully
        "401":
          description: Unauthorized access
        "404" :
          description: Entry not found
        "500":
          description: Internal server error
  /api/user/entry/reviewer/remove/{id}:
    post:
      tags: 
        - Entry
      description: >
        This endpoint removes a reviewer from an entry. If the entry is not found, it returns an error message.
      summary: remove a reviewer by user
      parameters:
        - in: path
          name: Id
          schema:
            type: string
          required: true
          description: The entry Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                  reviewer_id:
                    type: string
                    example: 5f7b1b3b7f3b9b0017b3b9b0
      responses:
        "200":
          description: reviewer removed successfully
        "401":
          description: Unauthorized access
        "404" :
          description: Entry not found
        "500":
          description: Internal server error
  /api/user/entry/delete/{id}:
    post:
      tags: 
        - Entry
      summary: delete a telecunsultation entry within 24 hours(after 24 hours cannot delete)
      description: >
        This endpoint deletes a Entry. If the entry is not found, it returns an error message.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The entry Id
      responses:
        "200":
          description: telecunsultation entry deleted successfully
        "401":
          description: Unauthorized access
        "404" :
          description: Entry not found
        "500":
          description: Internal server error
  /api/user/entry/shared/all:
    get:
      tags:
        - Entry
      summary: Get all shared entries
      description: >
        This endpoint returns a list of all shared teleconsultation entries. It supports pagination and search.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: page number for pagination
        - in: query
          name: filter
          schema:
            type: string
            enum: [All,Reviewed,Unreviewed]
          description: filter for entries
      responses:
        "200":
          description: A list of shared telecunsultation entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeleConEntry"
        "401":
          description: Unauthorized access
        "500":
          description: Internal server error
  /api/user/entry/shared/{id}:
    get:
      tags:
        - Entry
      summary: Get one shared entry (view only)
      description: >
        This endpoint returns the details of a shared Entry. If the entry is not found, it returns an error message.
      parameters:
        - name: id
          in: path
          required: true
          description: Entry ID
          schema:
            type: string
      responses:
        200:
          description: Entry details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeleConEntry'
        401:
          description: Unauthorized access
        404:
          description: Entry not found
        500:
          description: Internal Server Error
  /api/user/entry/shared/data/{id}:
    get:
      tags:
        - Entry
      summary: Get assigned entry details
      description: >
        This endpoint returns the details of a shared Entry with assignment data. If the entry is not found, it returns an error message.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the assignment
      responses:
        200:
          description: Detailed Entry with assignment data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeleConEntryWithAssignment'
        401:
          description: Unauthorized access
        404:
          description: Entry not found
        500:
          description: Internal Server Error
  /api/user/entry/reviews/{id}:
    get:
      summary: Get reviews of a Entry
      tags:
        - Entry
      description:
        This endpoint returns a list of reviews for a specific Entry. If no reviews are found, it returns an error message.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the Entry
      responses:
        200:
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        401:
          description: Unauthorized access
        404:
          description: Entry not found
        500:
          description: Internal Server Error
  /api/user/entry/reviewer/change/{id}:
    post:
      summary: Change the reviewer of an assignment
      tags:
        - Entry
      description:
        This endpoint allows users to change the reviewer of an assignment. If the assignment is not found, it returns an error message.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reviewer_id:
                  type: string
                  description: The ID of the new reviewer
      responses:
        200:
          description: Reviewer assigned successfully
        401:
          description: Unauthorized access
        404:
          description: Entry not found
        500:
          description: Internal Server Error
  /api/user/entry/review/{id}:
    post:
      summary: Add a new review to a Entry
      tags:
        - Entry
      description:
        This endpoint allows users to add a new review to a Entry. If the entry is not found, it returns an error message.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the Entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provisional_diagnosis:
                  type: string
                management_suggestions:
                  type: string
                referral_suggestions:
                  type: string
                other_comments:
                  type: string
      responses:
        200:
          description: Review added successfully
        401:
          description: Unauthorized access
        404:
          description: Entry not found
        500:
          description: Internal Server Error
  /api/user/entry/mark/{id}:
    post:
      summary: Mark an assignment as read
      tags:
        - Entry
      description:
        This endpoint allows users to mark an assignment as read. If the assignment is not found, it returns an error message.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the assignment
      responses:
        200:
          description: Marked as read
        401:
          description: Unauthorized access
        500:
          description: Internal Server Error
  /api/user/entry/open/{id}:
    post:
      summary: Mark a Entry as read
      tags:
        - Entry
      description:
        This endpoint allows users to mark a Entry as read. If the entry is not found, it returns an error message.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the Entry
      responses:
        200:
          description: Marked as read
        401:
          description: Unauthorized access
        500:
          description: Internal Server Error
...
components:
  schemas:
...
    TeleConEntry:
      type: object
      properties:
        patient:
          type: string
          description: ObjectId reference of the patient
        clinician_id:
          type: string
          description: ObjectId reference to the user
        complaint:
          type: string
          description: complaint of the patient
        start_time:
          type: string
          format: date-time
          description: start time of the telecon
        end_time:
          type: string
          format: date-time
          description: end time of the telecon
        findings:
          type: string
          description: findings of the telecon
        curent_habits:
          type: array
          items:
            type: string
          description: current habits of the patient
          default: []
        updated:
          type: boolean
          description: whether the telecon is updated
        reviewers:
          type: array
          items:
            type: string
            description: ObjectId reference to the User (reviewers)
        reviews:
          type: array
          items:
            type: string
            description: ObjectId reference to the Review
        images:
          type: array
          items:
            type: string
            description: ObjectId reference to the Image
        reports:
          type: array
          items: 
            type: string
            description: ObjectId reference to the Report
      required:
        - patient
        - clinician_id
    Review:
      type: object
      properties:
        id:
          type: string
        entry_id:
          type: string
        reviewer_id:
          type: string
        review_text:
          type: string
        review_date:
          type: string
          format: date-time
    TeleConEntryWithAssignment:
      allOf:
        - $ref: '#/components/schemas/TeleConEntry'
          properties:
            assignedAt:
              type: string
              format : date-time
            checked:
              type: boolean
            reviewed:
              type: boolean    
...